+++ src/nsidc/metgen/metgen.py
@@ -179,6 +179,11 @@ def init_config(configuration_file):
         "reference_file_regex",
         Prompt.ask("Reference science file regex"),
     )
+    cfg_parser.set(
+        constants.COLLECTION_SECTION_NAME,
+        "force_single_file_granules",
+        Prompt.ask("Force one data file per granule?", default=False),
+    )
     print()
 
     print()
@@ -495,6 +500,7 @@ def grouped_granule_files(configuration: config.Config) -> list[tuple]:
             file_list,
             premet_file_list,
             spatial_file_list,
+            configuration.force_single_file_granules,
         )
         for granule_key in granule_keys(configuration, file_list)
     ]
@@ -554,6 +560,7 @@ def granule_tuple(
     file_list: list,
     premet_list: list,
     spatial_list: list,
+    force_single_file_granules: bool = False,
 ) -> tuple:
     """
     Important! granule_regex argument must include a captured match group.
@@ -574,8 +581,17 @@ def granule_tuple(
         if re.search(granule_key, file.name) and re.search(browse_regex, file.name)
     }
 
+    # Match the file name exactly if the flag is set to force all data files to
+    # be treated as a separate granule. Otherwise, a substring match of the filename
+    # is ok.
+    match_func = re.fullmatch if force_single_file_granules else re.search
     data_file_paths = {
-        str(file) for file in file_list if re.search(granule_key, file.name)
+        # granule_key re.search here is the problem
+        # if set to single file granule, use re.fullmatch
+        # apply match_mode as selected above
+        str(file)
+        for file in file_list
+        if match_func(granule_key, file.name)
     } - browse_file_paths
 
     return (